fsoftuni/requirements.txt
aioredis==2.0.1
anyascii==0.3.1
arabic-reshaper==2.1.3
asgi-redis
asgiref
asn1crypto==1.5.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
beautifulsoup4==4.9.3
biopython==1.79
boto3==1.24.13
botocore==1.27.13
certifi==2022.6.15
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==2.0.12
click==8.1.3
commonmark==0.9.1
constantly==15.1.0
cryptography==38.0.3
cssselect2==0.6.0
cycler==0.11.0
daphne==4.0.0
Deprecated==1.2.13
Django==4.0.8
django-filter==21.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-reset-migrations==0.4.0
django-storages==1.12.3
django-taggit==2.1.0
django-treebeard==4.5.1
django-utils2==2.12.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
docx2pdf==0.1.8
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
exceptiongroup==1.0.0
fonttools==4.37.1
future==0.18.2
greenlet==1.1.2
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
hyperlink==21.0.0
idna==3.3
incremental==22.10.0
jmespath==1.0.1
kiwisolver==1.4.4
l18n==2021.3
lxml==4.9.1
matplotlib==3.5.3
mock==1.0.1
msgpack==1.0.4
msgpack-python==0.5.6
numpy==1.23.0
opencv-python==4.6.0.66
openpyxl==3.0.10
oscrypto==1.3.0
outcome==1.2.0
packaging==21.3
pandas==1.4.3
pdfkit==1.0.0
Pillow==9.1.1
polib==1.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
pyOpenSSL==22.1.0
pyparsing==3.0.9
PyPDF3==1.0.6
pysam==0.19.1
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.6
python-docx==0.8.11
python-utils==3.3.3
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
redis==4.2.0
regex==2022.10.31
reportlab==3.6.11
requests==2.28.0
rich==12.4.4
RUST==0.1.1
s3transfer==0.6.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
sqlparse==0.4.2
svglib==1.4.1
tablib==3.2.1
telepath==0.2
tinycss2==1.1.1
tqdm==4.64.0
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
unoconv==0.9.0
uritools==4.0.0
urllib3==1.26.9
wagtail==3.0.2
wagtail-localize==1.2.1
wagtail-modeladmin-sortable==0.0.2
webencodings==0.5.1
websocket-client==1.4.1
websockets==10.4
Willow==1.4.1
wrapt==1.14.1
wsproto==1.2.0
xhtml2pdf==0.2.8
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
zope.interface==5.5.1
